FIX: Move chat service worker source out of `assets/javascripts/`
`Plugin::JsManager#compile_js_bundle` globs every file under
`plugins/<name>/assets/javascripts/` into the plugin's main JS bundle,
which is loaded as part of the regular Ember app (where `self ===
window`). Service worker source files placed there call
`self.registerNotificationActionHandler(...)`, which is undefined in
that context — throwing on load and breaking the entire chat
frontend. That cascade was timing out every chat system test that
visits a channel and the cross-plugin "plugin javascript assets" test.

Move the file to `plugins/chat/assets/service-worker/` and update the
`register_service_worker` registration. The `service-worker.js.erb`
template reads it via `File.read` regardless of location under
`assets/`, so the runtime behavior is unchanged.
