+++ plugins/chat/assets/service-worker/chat-service-worker-extensions.js
